Feat: adding new input and remove input on Pemeriksaan Proses Packing

+++ app/Http/Controllers/PackingController.php
@@ -9,7 +9,6 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
-use Carbon\Carbon;
 use Intervention\Image\ImageManager;
 use Intervention\Image\Drivers\Gd\Driver;
 use TCPDF;
@@ -70,64 +69,53 @@ class PackingController extends Controller
         : 'Produksi RTT';
 
         $request->validate([
-            'date'        => 'required|date',
-            'shift'       => 'required',
-            'nama_produk' => 'required',
-            'waktu'       => 'required',
-            'kalibrasi'   => 'required',
-            'qrcode'      => 'required',
-
+            'date'          => 'required|date',
+            'shift'         => 'required',
+            'nama_produk'   => 'required',
+            'waktu'         => 'required',
+            'kalibrasi'     => 'nullable|string',
+            'qrcode'        => 'nullable|image|mimes:jpeg,png,jpg|max:2048',
             'kode_printing' => 'nullable',
-
-            'kode_toples'   => 'nullable|string',
+            'kode_toples'   => 'required|string',
             'kode_karton'   => 'nullable|string',
             'suhu'          => 'nullable|numeric',
             'speed'         => 'nullable|numeric',
             'kondisi_segel' => 'nullable|string',
-            'berat_toples'  => 'nullable|numeric',
-            'berat_pouch'   => 'nullable|numeric',
-            'no_lot'        => 'nullable|string',
-            'tgl_kedatangan'=> 'nullable|date',
-            'nama_supplier' => 'nullable|string',
-            'keterangan'    => 'nullable|string',
+            'jumlah_produk' => 'nullable|integer', 
+            'berat_pcs'     => 'nullable|numeric', 
+            'berat_pack'    => 'nullable|numeric', 
+            
+            // Validasi Input Dinamis Array Kemasan
+            'data_kemasan'                  => 'nullable|array',
+            'data_kemasan.*.jenis_kemasan'  => 'required|string',
+            'data_kemasan.*.no_lot_kemasan' => 'required|string',
+            'data_kemasan.*.tgl_kedatangan' => 'nullable|date',
+            'data_kemasan.*.nama_supplier'  => 'nullable|string',
+            'keterangan'                    => 'nullable|string',
         ]);
 
-        $kodePrintingFinal = null;
-        if ($request->hasFile('kode_printing')) {
-
-            $file = $request->file('kode_printing');
-
-            if ($file->getSize() > 2 * 1024 * 1024) {
-                return back()
-                ->withErrors(['kode_printing' => 'Ukuran file maksimal 2 MB.'])
-                ->withInput();
-            }
-
-            $kodePrintingFinal = $this->compressAndStore($file, 'printing');
-
-        } else {
-            $kodePrintingFinal = $request->kode_printing;
-        }
+        $kodePrintingFinal = $request->hasFile('kode_printing') ? $this->compressAndStore($request->file('kode_printing'), 'printing') : null;
+        $qrcodeFinal = $request->hasFile('qrcode') ? $this->compressAndStore($request->file('qrcode'), 'qrcode') : null;
 
         Packing::create([
             'date'                => $request->date,
             'shift'               => $request->shift,
             'nama_produk'         => $request->nama_produk,
             'waktu'               => $request->waktu,
-            'kalibrasi'           => $request->kalibrasi,
-            'qrcode'              => $request->qrcode,
+            'kalibrasi'           => $request->has('kalibrasi') ? 'Ok' : 'Tidak Ok',
+            'qrcode'              => $qrcodeFinal,
             'kode_printing'       => $kodePrintingFinal,
-
             'kode_toples'         => $request->kode_toples,
             'kode_karton'         => $request->kode_karton,
             'suhu'                => $request->suhu,
             'speed'               => $request->speed,
             'kondisi_segel'       => $request->kondisi_segel,
-            'berat_toples'        => $request->berat_toples,
-            'berat_pouch'         => $request->berat_pouch,
-            'no_lot'              => $request->no_lot,
-            'tgl_kedatangan'      => $request->tgl_kedatangan,
-            'nama_supplier'       => $request->nama_supplier,
+            'jumlah_produk'       => $request->jumlah_produk,
+            'berat_pcs'           => $request->berat_pcs,
+            'berat_pack'          => $request->berat_pack,
+            
+            // Encode menjadi JSON string sebelum masuk DB
+            'data_kemasan'        => $request->has('data_kemasan') ? json_encode($request->data_kemasan) : null,
             'keterangan'          => $request->keterangan,
 
             'username'            => $username,
@@ -138,8 +126,7 @@ class PackingController extends Controller
             'status_spv'          => "0",
         ]);
 
-        return redirect()->route('packing.index')
-        ->with('success', 'Pemeriksaan Proses Cartoning berhasil disimpan');
+        return redirect()->route('packing.index')->with('success', 'Pemeriksaan Proses Packing berhasil disimpan');
     }
 
     public function update(string $uuid)
@@ -161,80 +148,43 @@ class PackingController extends Controller
         $username_updated = Auth::user()->username ?? 'User RTM';
 
         $request->validate([
-            'date'        => 'required|date',
-            'shift'       => 'required',
-            'nama_produk' => 'required',
-            'waktu'       => 'required',
-            'kalibrasi'   => 'required',
-            'qrcode'      => 'required',
-            'kode_printing' => 'nullable',
-
-            'kode_toples'   => 'nullable|string',
-            'kode_karton'   => 'nullable|string',
-            'suhu'          => 'nullable|numeric',
-            'speed'         => 'nullable|numeric',
-            'kondisi_segel' => 'nullable|string',
-            'berat_toples'  => 'nullable|numeric',
-            'berat_pouch'   => 'nullable|numeric',
-            'no_lot'        => 'nullable|string',
-            'tgl_kedatangan'=> 'nullable|date',
-            'nama_supplier' => 'nullable|string',
+            'date'          => 'required|date',
+            'shift'         => 'required',
+            'nama_produk'   => 'required',
+            'waktu'         => 'required',
+            'kalibrasi'     => 'nullable|string',
+            'qrcode'        => 'nullable|image|mimes:jpeg,png,jpg|max:2048',
+            'kode_toples'   => 'required|string',
+            'data_kemasan'  => 'nullable|array',
             'keterangan'    => 'nullable|string',
         ]);
 
-        $kodePrintingFinal = $packing->kode_printing;
-
-        if ($request->hasFile('kode_printing')) {
-
-            $file = $request->file('kode_printing');
-
-            if ($file->getSize() > 2 * 1024 * 1024) {
-                return back()
-                ->withErrors(['kode_printing' => 'Ukuran file maksimal 2 MB.'])
-                ->withInput();
-            }
-            if (!empty($packing->kode_printing) && str_contains($packing->kode_printing, 'storage/')) {
-                $oldPath = str_replace('storage/', 'public/', $packing->kode_printing);
-
-                if (Storage::exists($oldPath)) {
-                    Storage::delete($oldPath);
-                }
-            }
-
-            $kodePrintingFinal = $this->compressAndStore($file, 'printing');
-
-        } else {
-            if (is_string($request->kode_printing)) {
-                $kodePrintingFinal = $request->kode_printing;
-            }
-        }
+        // File Handler Code (Compressed)
+        $kodePrintingFinal = $request->hasFile('kode_printing') ? $this->compressAndStore($request->file('kode_printing'), 'printing') : $packing->kode_printing;
+        $qrcodeFinal = $request->hasFile('qrcode') ? $this->compressAndStore($request->file('qrcode'), 'qrcode') : $packing->qrcode;
 
         $packing->update([
             'date'                => $request->date,
             'shift'               => $request->shift,
             'nama_produk'         => $request->nama_produk,
             'waktu'               => $request->waktu,
-            'kalibrasi'           => $request->kalibrasi,
-            'qrcode'              => $request->qrcode,
+            'kalibrasi'           => $request->has('kalibrasi') ? 'Ok' : 'Tidak Ok',
+            'qrcode'              => $qrcodeFinal,
             'kode_printing'       => $kodePrintingFinal,
-
             'kode_toples'         => $request->kode_toples,
             'kode_karton'         => $request->kode_karton,
             'suhu'                => $request->suhu,
             'speed'               => $request->speed,
             'kondisi_segel'       => $request->kondisi_segel,
-            'berat_toples'        => $request->berat_toples,
-            'berat_pouch'         => $request->berat_pouch,
-            'no_lot'              => $request->no_lot,
-            'tgl_kedatangan'      => $request->tgl_kedatangan,
-            'nama_supplier'       => $request->nama_supplier,
+            'jumlah_produk'       => $request->jumlah_produk,
+            'berat_pcs'           => $request->berat_pcs,
+            'berat_pack'          => $request->berat_pack,
+            'data_kemasan'        => $request->has('data_kemasan') ? json_encode($request->data_kemasan) : null,
             'keterangan'          => $request->keterangan,
-
             'username_updated'    => $username_updated,
         ]);
 
-        return redirect()->route('packing.index')
-        ->with('success', 'Pemeriksaan Proses Cartoning berhasil diperbarui');
+        return redirect()->route('packing.index')->with('success', 'Pemeriksaan Proses Packing berhasil diperbarui');
     }
 
     public function edit(string $uuid)
@@ -255,78 +205,34 @@ class PackingController extends Controller
         $packing = Packing::where('uuid', $uuid)->firstOrFail();
 
         $request->validate([
-            'date'        => 'required|date',
-            'shift'       => 'required',
-            'nama_produk' => 'required',
-            'waktu'       => 'required',
-            'kalibrasi'   => 'required',
-            'qrcode'      => 'required',
-            'kode_printing' => 'nullable',
-
-            'kode_toples'   => 'nullable|string',
-            'kode_karton'   => 'nullable|string',
-            'suhu'          => 'nullable|numeric',
-            'speed'         => 'nullable|numeric',
-            'kondisi_segel' => 'nullable|string',
-            'berat_toples'  => 'nullable|numeric',
-            'berat_pouch'   => 'nullable|numeric',
-            'no_lot'        => 'nullable|string',
-            'tgl_kedatangan'=> 'nullable|date',
-            'nama_supplier' => 'nullable|string',
+            'date'          => 'required|date',
+            'shift'         => 'required',
+            'nama_produk'   => 'required',
+            'waktu'         => 'required',
+            'kode_toples'   => 'required|string',
+            'data_kemasan'  => 'nullable|array',
             'keterangan'    => 'nullable|string',
         ]);
-
-        $kodePrintingFinal = $packing->kode_printing;
-
-        if ($request->hasFile('kode_printing')) {
-
-            $file = $request->file('kode_printing');
-
-            if ($file->getSize() > 2 * 1024 * 1024) {
-                return back()
-                ->withErrors(['kode_printing' => 'Ukuran file maksimal 2 MB.'])
-                ->withInput();
-            }
-            if (!empty($packing->kode_printing) && str_contains($packing->kode_printing, 'storage/')) {
-                $oldPath = str_replace('storage/', 'public/', $packing->kode_printing);
-
-                if (Storage::exists($oldPath)) {
-                    Storage::delete($oldPath);
-                }
-            }
-
-            $kodePrintingFinal = $this->compressAndStore($file, 'printing');
-
-        } else {
-            if (is_string($request->kode_printing)) {
-                $kodePrintingFinal = $request->kode_printing;
-            }
-        }
 
         $packing->update([
             'date'                => $request->date,
             'shift'               => $request->shift,
             'nama_produk'         => $request->nama_produk,
             'waktu'               => $request->waktu,
-            'kalibrasi'           => $request->kalibrasi,
-            'qrcode'              => $request->qrcode,
-            'kode_printing'       => $kodePrintingFinal,
-
+            'kalibrasi'           => $request->has('kalibrasi') ? 'Ok' : 'Tidak Ok',
             'kode_toples'         => $request->kode_toples,
             'kode_karton'         => $request->kode_karton,
             'suhu'                => $request->suhu,
             'speed'               => $request->speed,
             'kondisi_segel'       => $request->kondisi_segel,
-            'berat_toples'        => $request->berat_toples,
-            'berat_pouch'         => $request->berat_pouch,
-            'no_lot'              => $request->no_lot,
-            'tgl_kedatangan'      => $request->tgl_kedatangan,
-            'nama_supplier'       => $request->nama_supplier,
+            'jumlah_produk'       => $request->jumlah_produk,
+            'berat_pcs'           => $request->berat_pcs,
+            'berat_pack'          => $request->berat_pack,
+            'data_kemasan'        => $request->has('data_kemasan') ? json_encode($request->data_kemasan) : null,
             'keterangan'          => $request->keterangan,
         ]);
 
-        return redirect()->route('packing.index')
-        ->with('success', 'Pemeriksaan Proses Cartoning berhasil diperbarui');
+        return redirect()->route('packing.index')->with('success', 'Pemeriksaan Proses Packing berhasil diperbarui');
     }
 
     public function verification(Request $request)
@@ -370,57 +276,48 @@ class PackingController extends Controller
             'tgl_update_spv'  => now(),
         ]);
 
-        return redirect()->route('packing.index')
-        ->with('success', 'Status Verifikasi Pemeriksaan Proses Cartoning berhasil diperbarui.');
+        return redirect()->route('packing.index')->with('success', 'Status Verifikasi Pemeriksaan Proses Packing berhasil diperbarui.');
     }
 
     public function destroy($uuid)
     {
         $packing = Packing::where('uuid', $uuid)->firstOrFail();
 
-        // Hapus file gambar jika ada
-        $proses = json_decode($packing->proses, true) ?? [];
-        foreach ($proses as $row) {
-            if (!empty($row['kode_printing'])) {
-                $path = str_replace('storage/', 'public/', $row['kode_printing']);
-                if (Storage::exists($path)) {
-                    Storage::delete($path);
-                }
-            }
+        if (!empty($packing->kode_printing)) {
+            $path = str_replace('storage/', 'public/', $packing->kode_printing);
+            if (Storage::exists($path)) { Storage::delete($path); }
+        }
+
+        if (!empty($packing->qrcode)) {
+            $pathQr = str_replace('storage/', 'public/', $packing->qrcode);
+            if (Storage::exists($pathQr)) { Storage::delete($pathQr); }
         }
 
         $packing->delete();
-
-        return redirect()->route('packing.index')
-        ->with('success', 'Pemeriksaan Proses Packing berhasil dihapus');
+        return redirect()->route('packing.index')->with('success', 'Pemeriksaan Proses Packing berhasil dihapus');
     }
 
     public function recyclebin()
     {
-        $packing = Packing::onlyTrashed()
-        ->orderBy('deleted_at', 'desc')
-        ->paginate(10);
-
+        $packing = Packing::onlyTrashed()->orderBy('deleted_at', 'desc')->paginate(10);
         return view('form.packing.recyclebin', compact('packing'));
     }
+
     public function restore($uuid)
     {
         $packing = Packing::onlyTrashed()->where('uuid', $uuid)->firstOrFail();
         $packing->restore();
-
-        return redirect()->route('packing.recyclebin')
-        ->with('success', 'Data berhasil direstore.');
+        return redirect()->route('packing.recyclebin')->with('success', 'Data berhasil direstore.');
     }
+
     public function deletePermanent($uuid)
     {
         $packing = Packing::onlyTrashed()->where('uuid', $uuid)->firstOrFail();
         $packing->forceDelete();
-
-        return redirect()->route('packing.recyclebin')
-        ->with('success', 'Data berhasil dihapus permanen.');
+        return redirect()->route('packing.recyclebin')->with('success', 'Data berhasil dihapus permanen.');
     }
 
-    public function exportPdf(Request $request)
+        public function exportPdf(Request $request)
     {
         $date = $request->input('date');
         $shift = $request->input('shift');
@@ -453,8 +350,8 @@ class PackingController extends Controller
         // Set document information
         $pdf->SetCreator(PDF_CREATOR);
         $pdf->SetAuthor('Your Name/Company');
-        $pdf->SetTitle('Pemeriksaan Proses Cartoning');
-        $pdf->SetSubject('Pemeriksaan Proses Cartoning');
+        $pdf->SetTitle('Pemeriksaan Proses Packing');
+        $pdf->SetSubject('Pemeriksaan Proses Packing');
 
         $pdf->SetPrintHeader(false);
         $pdf->SetPrintFooter(false);
@@ -496,26 +393,18 @@ class PackingController extends Controller
 
         exit();
     }
-
-    /**
-     * 🧰 Helper untuk kompres dan simpan gambar
-     */
+    
     private function compressAndStore($file, $prefix)
     {
         $manager = new ImageManager(new Driver());
-        $path = 'public/uploads/packing'; // ✅ pindah ke folder aman
+        $path = 'public/uploads/packing';
         $filename = $prefix . '_' . Str::uuid() . '.jpg';
 
-        // Pastikan folder ada
-        if (!Storage::exists($path)) {
-            Storage::makeDirectory($path, 0755, true);
-        }
+        if (!Storage::exists($path)) { Storage::makeDirectory($path, 0755, true); }
 
-        // Resize dan kompres gambar
         $image = $manager->read($file)->scale(width: 1280)->toJpeg(quality: 75);
         Storage::put("$path/$filename", (string) $image);
 
-        // Return path yang bisa dipakai untuk view
         return "storage/uploads/packing/$filename";
     }
 }
